Add authentication checks and redirect users to login page if not logged in

+++ src/components/WorkoutDetails.js
@@ -1,18 +1,30 @@
 import { useWorkoutsContext} from '../hooks/useWorkoutsContext'
 import  formatDistanceToNow  from "date-fns/formatDistanceToNow";
+import {useAuthContext} from '../hooks/useAuthContext'
 
 function WorkoutDetails({workout}) {
 
   const {dispatch}= useWorkoutsContext()
+  const {user} = useAuthContext()
+
   const handleClick=async()=>{
+    if(!user){
+      return
+    }
+
     const response = await fetch('/api/workouts/'+ workout._id,{
       method:'DELETE',
+      headers:{
+        'Authorization' : `Bearer ${user.token}`
+      }
     })
+
     const json =await response.json()
 
     if(response.ok){
       dispatch({type: 'DELETE_WORKOUT',payload:json})
     }
+
   }
   return (
     <div className="workout-details">
